refactored name of settings objects to algorithmsettings objects
